+++ startup.utils.js
@@ -1,11 +1,12 @@
 // A sample extension of the $startup.utils object
-(function(S){
-	if (!S) { return false; }
+(function($tartup){
+   
+	if (!$tartup) { return false; }
 	
-	var Add = S.utils.Add;
+	var AddUtil = $tartup.addUtil;
 	
 	//Gets query string values
-	Add("queryStr",function(param){
+	AddUtil("queryStr",function(param){
 		var x, val = false, query = window.location.search.substr(1), qArr = query.split("&"), len = qArr.length;
 		for (x = 0; x < len; x++) {
 			var pair = qArr[x].split("=");
@@ -15,7 +16,7 @@
 	});
 	
 	// Get hash values
-	Add("getHash",function(param){
+	AddUtil("getHash",function(param){
 		var x, val = false, query = window.location.hash.substr(1), qArr = query.split("&"), len = qArr.length;
 		for (x = 0; x < len; x++) {
 			var pair = qArr[x].split("=");
@@ -24,23 +25,18 @@
 		return val;
 	});
 	
-	// console log
-	Add("log",function(str){
-		return log(str);
-	});
-	
-	//utils for date/time items
+	// a simple util for date/time items
 	/*	date/time formatting
 		Sample usage:
 			var dateObj = new Date();
 			$tartup.utils.datetime.format(dateObj,"Today is: ${ddTH} of ${month} ${yr}");
-			- will return string: Today is 9th of December 2009
+			- will return string e.g.: Today is 9th of December 2009
 	*/
-	
-	Add("datetime",{
-		format: function(date,formatStr) {
-			formatStr = formatStr||null; // format is optional
-		
+	AddUtil("datetime",{
+	   format: function(date,formatStr) {
+	      formatStr = formatStr||null; // format is optional
+         var i;
+         
 			var obj = {
 				// date
 				month: date.getUTCMonth(),
@@ -54,72 +50,91 @@
 				ampm: ((date.getHours()<12)?'am':'pm'),
 				fullStr: date.toString()
 			};
-		
+			
+			// old english(?) format
+			var oldEnglish = function(n) {
+			   n = n || "";
+			   n = n.toString();
+			   
+			   if (n.length < 1) { return n; }
+			   
+   			var s = "", lastDigit = parseInt(n.charAt(n.length-1),10);
+   			switch (lastDigit) {
+   				case 1: s = "st";break;
+   				case 2: s = "nd";break;
+   				case 3: s = "rd";break;
+   				default: s = "th";break;
+   			}
+   			
+   			return (n + s);
+			};
+			
+			// Return month string
+   		var getUTCMonthString = function(n,type) {
+   			type = type||'full'; //type is optional
+   			var month = ["January","February","March","April","May","June","July","August","September","October","November","December"], ret = month[n] || "";
+   			
+   			if (type==="short") {
+   			   ret = ret.slice(0,3);
+   			}
+   			return ret;
+   		};
+   		
+   		// to string
+   		var toStr = function(val) {
+   		   val = val || "";
+   		   return val.toString();
+   		};
+   		
+   		// Return week days string
+   		var getDayString = function(n,type) {
+   			type = type || "full";
+   			var day = ["Sunday","Monday","Tuesday","Wednesday","Thursday","Friday","Saturday"], ret = day[n];
+   			
+   			if (type==="short") {
+   			   ret = ret.slice(0,3);
+   			}
+   			return ret;
+   		};
+         
+         // replacements
 			var replace = {
 				"${mm}": (obj.month+1),
-				"${month}": this.getUTCMonthString(obj.month),
-				"${month:short}": this.getUTCMonthString(obj.month,'short'),
-				"${day}": this.getDayString(obj.day),
-				"${day:short}": this.getDayString(obj.day,'short'),
+				"${month}": getUTCMonthString(obj.month),
+				"${month:short}": getUTCMonthString(obj.month,"short"),
+				"${day}": getDayString(obj.day),
+				"${day:short}": getDayString(obj.day,"short"),
 				"${yr}": obj.year,
-				"${dd}": ((obj.date<10)?'0'+obj.date:obj.date),
-				"${ddTH}": this.oldEnglish(obj.date),
+				"${yr:short}": toStr(obj.year).substr(2),
+				"${dd}": ((obj.date<10) ? ("0" + toStr(obj.date)) : toStr(obj.date)),
+				"${ddTH}": oldEnglish(obj.date),
 				"${hr}": ((obj.hour>12)?obj.hour-12:obj.hour),
 				"${hr:mil}": obj.hour,
-				"${min}": ((obj.minutes<10)?'0'+obj.minutes:obj.minutes),
-				"${sec}": ((obj.seconds<10)?'0'+obj.seconds:obj.seconds),
+				"${min}": ((obj.minutes<10) ? ("0" + toStr(obj.minutes)) : toStr(obj.minutes)),
+				"${sec}": ((obj.seconds<10) ? ("0" + toStr(obj.seconds)) : toStr(obj.seconds)),
 				"${ampm}": obj.ampm,
 				"${AMPM}": (obj.ampm).toUpperCase(),
 				"${AmPm}": ((obj.ampm).charAt(0)).toUpperCase()+(obj.ampm).charAt(1)
 			};
-		
-			if (formatStr===null) { // from old format date
-				var month = this.getUTCMonthString(obj.month), dateString = date.toDateString();
-				return month + ' ' + dateString.substr(8).replace(' ', ', '); //Wed Sep 02 2009
-			} else {
-				var i; for (i in replace) {
-					var idx = formatStr.indexOf(i);
-					while (idx!==-1) {
-						formatStr = formatStr.replace(i,replace[i]);
-						idx = formatStr.indexOf(i);
-					}
+
+			for (i in replace) {
+				var idx = formatStr.indexOf(i);
+				while (idx!==-1) {
+					formatStr = formatStr.replace(i,replace[i]);
+					idx = formatStr.indexOf(i);
 				}
-				return formatStr;
 			}
-		},
+			return formatStr;
+	   }
+	});
+
+	var log = function(str) {
+		if (window.console && console.log) { console.log(str); }
+	};
 	
-		// old english(?) format
-		oldEnglish: function(num) {
-			num+="";
-			var lastDigit = parseInt(num.charAt(num.length-1),10);
-			switch (lastDigit) {
-				case 1: num+="st";break;
-				case 2: num+="nd";break;
-				case 3: num+="rd";break;
-				default: num+="th";break;
-			}
-			return num;
-		},
-	
-		// Return month string
-		getUTCMonthString: function(monthNum,type) {
-			type = type||'full'; //type is optional
-			var month = ["January","February","March","April","May","June","July","August","September","October","November","December"],
-				sel = month[monthNum];
-			return (type==="short") ? sel.substr(0,3) : sel;
-		},
-	
-		// Return week days string
-		getDayString: function(dayNum,type) {
-			type = type||'full';
-			var day = ["Sunday","Monday","Tuesday","Wednesday","Thursday","Friday","Saturday"],
-				sel = day[dayNum];
-			return (type==="short") ? sel.substr(0,3) : sel;	
-		}
+	// console log
+	AddUtil("log",function(str){
+		return log(str);
 	});
 	
-	function log(str) {
-		if (window.console && console.log) { console.log(str); }
-	}
-	
-}($tartup || false));
+}(window.$tartup));